+++ NEWS.md
@@ -1,3 +1,5 @@
+# rxode2 5.1.7
+
 # rxode2 5.1.6
 
 ## New features